fix(cheatsheet): reuse the cached catalog, and stop losing diagnostics
Second batch from the ECC audit.

- CheatsheetManager resolved the note catalog itself through IEntityLookup,
  uncached, on every read and write, while the serializer separately did its own
  (blocking, synchronous) fetch of the same value - two round-trips per call.
  IEntityLookup served nothing else here, so it is replaced by the existing
  ICheatsheetCatalogProvider, which caches on success only, so a transient
  failure still retries rather than being remembered as "missing".

- CheatsheetAppStartupFilter treated a FAILED catalog lookup the same as "no
  catalog yet" and fell through to creating one, discarding the real error. A
  lookup failure now logs the cause and skips seeding.

- ReadInt accepted any JSON number and narrowed it with a cast, so schemaVersion
  1.7 was silently stored as 1. It was the only reader here that did not require
  an exact type match; it now uses TryGetInt32 and leaves anything else to
  ExtraFields like every other field.

- LoadCardsAsync dropped unreadable notes from list results with no log at all.
  Dropping them is correct - one bad note must not take the wallet down - but
  the card disappears from the user's view, so it now logs the note id and
  subject. The manager takes an optional ILogger.

Tests: the validator's author lookup answered identically for any id, so a
validator checking a hardcoded author instead of card.AuthorId passed unchanged;
it is now scoped to the expected id with a test that proves the id flows from
the card. Four validator failure tests asserted only that validation failed, not
which rule fired, so a mis-attributed message went unnoticed.

Mutation-verified: reverting ReadInt to the truncating form fails the new
fractional-schemaVersion test, and hardcoding the validator's author id fails
the new author-flow test.

+++ src/Hmm.Cheatsheet/NoteSerialize/CheatsheetJsonNoteSerialize.cs
@@ -435,12 +435,16 @@ namespace Hmm.Cheatsheet.NoteSerialize
 
         private static int ReadInt(JsonElement element, string name, int defaultValue, HashSet<string> consumed)
         {
+            // TryGetInt32, not TryGetDouble-then-cast: the latter accepted 1.7 and
+            // silently narrowed it to 1, discarding what the client sent. Every
+            // other reader here consumes only on an exact type match and leaves
+            // anything else to ExtraFields; this one was the exception.
             if (element.TryGetProperty(name, out var property) &&
                 property.ValueKind == JsonValueKind.Number &&
-                property.TryGetDouble(out var value))
+                property.TryGetInt32(out var value))
             {
                 consumed.Add(name);
-                return (int)value;
+                return value;
             }
 
             return defaultValue;
